refactor: input component was separated in 3 new components
