+++ src/components/atom/Input.tsx
@@ -12,14 +12,8 @@ import {
 } from 'react';
 import mixpanel from 'mixpanel-browser';
 
-import {
-  MdAdd,
-  MdClose,
-  MdKeyboardArrowDown,
-  MdKeyboardArrowUp,
-  MdRemoveRedEye,
-} from 'react-icons/md';
-import { getGeocode, getLatLng, Suggestion } from 'use-places-autocomplete';
+import { MdRemoveRedEye } from 'react-icons/md';
+import { Suggestion } from 'use-places-autocomplete';
 import { EXAMS as EXAMS_CONSTANT } from '@/constants/examsSearch';
 
 import Exam from '@/@types/Exam';
@@ -29,10 +23,6 @@ import {
   UserInput,
   InputIcon,
   InputTextArea,
-  SuggestionArea,
-  SelectedExams,
-  SelectedExamsSummary,
-  SelectedExamsDetail,
   ErrorMessage,
 } from '@/styles/components/atom/Input';
 import useClickOutsideRef from '@/hooks/clickOutside';
@@ -40,7 +30,7 @@ import { useAuth } from '@/hooks/auth';
 
 import { useField } from '@unform/core';
 
-type SuggestionProps =
+export type SuggestionProps =
   | {
       type: 'exams';
       data: Exam[];
@@ -79,10 +69,8 @@ const Input = ({
   const [isFocused, setIsFocused] = useState(false);
   const [isFilled, setIsFilled] = useState(false);
   const [hasSuggestions, setHasSuggestions] = useState(false);
-  const [isOpenSelectedExams, setIsOpenSelectedExams] = useState(false);
   const [inputType, setInputType] = useState('text');
 
-  const { addAddress, addExam, exams, removeExam } = useSearchExam();
   const { user } = useAuth();
 
   useEffect(() => {
@@ -124,10 +112,6 @@ const Input = ({
     setHasSuggestions(false);
   }, [inputRef, getInputValue, suggestions, value]);
 
-  const handleClickOutsideSelectedExams = useCallback(() => {
-    setIsOpenSelectedExams(false);
-  }, []);
-
   const handleInputChange = useCallback(() => {
     setIsFilled(!!inputRef.current?.value);
 
@@ -136,81 +120,7 @@ const Input = ({
     suggestions && setHasSuggestions(true);
   }, [suggestions, getInputValue]);
 
-  const handlePlaceSelect = useCallback(
-    (address: string) => () => {
-      // When user selects a place, we can replace the keyword without request data from API
-      // by setting the second parameter as "false"
-      // setHasAddress(true);
-      suggestions.type === 'address' && suggestions.getSelectedAddress(address);
-
-      setHasSuggestions(false);
-
-      suggestions.type === 'address' && suggestions.clearSuggestions();
-
-      // Get latitude and longitude via utility functions
-      getGeocode({
-        address,
-      })
-        .then(placeResults => getLatLng(placeResults[0]))
-        .then(({ lat, lng }) => {
-          console.log('📍 Coordinates: ', { lat, lng });
-
-          addAddress({
-            address,
-            latitude: lat,
-            longitude: lng,
-          });
-
-          user && mixpanel.identify(user.id);
-          mixpanel.track('Add Address To Search', {
-            Address: address,
-            Latitude: lat,
-            Longitude: lng,
-          });
-        })
-        .catch(error => {
-          console.log('😱 Error: ', error);
-        });
-    },
-    [suggestions, user, addAddress],
-  );
-
-  const handleExamSelect = useCallback(
-    (exam: Exam) => {
-      addExam(exam);
-
-      setHasSuggestions(false);
-      suggestions.type === EXAMS_CONSTANT && suggestions.clearSuggestions;
-
-      inputRef.current.value = '';
-      setIsFilled(false);
-
-      user && mixpanel.identify(user.id);
-      mixpanel.track('Add Exam To Search', {
-        Exam: exam.title,
-      });
-    },
-    [suggestions, addExam, user],
-  );
-
-  const handleExamRemove = useCallback(
-    (exam: Exam) => {
-      exams.length === 0 && setIsOpenSelectedExams(false);
-
-      user && mixpanel.identify(user.id);
-      mixpanel.track('Remove Exam From Search', {
-        Exam: exam.title,
-      });
-
-      removeExam(exam.id);
-    },
-    [exams, removeExam, user],
-  );
-
   const clickOutsideUserInputRef = useClickOutsideRef(handleClickOutsideInput);
-  const clickOutsideUserSelectedExams = useClickOutsideRef(
-    handleClickOutsideSelectedExams,
-  );
 
   user && mixpanel.identify(user.id);
   mixpanel.track_links(
@@ -256,35 +166,6 @@ const Input = ({
         )}
       </UserInput>
 
-      {hasSuggestions && suggestions.type === EXAMS_CONSTANT && (
-        <SuggestionArea>
-          {suggestions.data.map(exam => (
-            <article key={exam.id} onClick={() => handleExamSelect(exam)}>
-              <p>{exam.title}</p>
-              <MdAdd />
-            </article>
-          ))}
-        </SuggestionArea>
-      )}
-
-      {hasSuggestions && suggestions.type === 'address' && (
-        <SuggestionArea>
-          {suggestions.data.map(suggestion => {
-            const {
-              place_id,
-              structured_formatting: { main_text, secondary_text },
-            } = suggestion;
-            const addressText = `${main_text} ${secondary_text}`;
-            return (
-              <article key={place_id} onClick={handlePlaceSelect(addressText)}>
-                <p>{addressText}</p>
-                <MdAdd />
-              </article>
-            );
-          })}
-        </SuggestionArea>
-      )}
-
       {error && (
         <ErrorMessage>
           <p>
@@ -304,41 +185,6 @@ const Input = ({
           </p>
         </ErrorMessage>
       )}
-
-      <SelectedExams ref={clickOutsideUserSelectedExams}>
-        {exams?.length > 0 && suggestions?.type === EXAMS_CONSTANT && (
-          <SelectedExamsSummary
-            onClick={() => setIsOpenSelectedExams(!isOpenSelectedExams)}
-          >
-            {exams.length === 1 ? (
-              <p>1 exame selecionado</p>
-            ) : (
-              <p>{exams.length} exames selecionados</p>
-            )}
-            <button
-              type="button"
-              onClick={() => setIsOpenSelectedExams(!isOpenSelectedExams)}
-            >
-              {isOpenSelectedExams ? (
-                <MdKeyboardArrowUp />
-              ) : (
-                <MdKeyboardArrowDown />
-              )}
-            </button>
-          </SelectedExamsSummary>
-        )}
-
-        {isOpenSelectedExams && (
-          <SelectedExamsDetail>
-            {exams.map(exam => (
-              <article key={exam.id}>
-                <p>{exam.title}</p>
-                <MdClose onClick={() => handleExamRemove(exam)} />
-              </article>
-            ))}
-          </SelectedExamsDetail>
-        )}
-      </SelectedExams>
     </InputContainer>
   );
 };
